Move files into packages (command and model)
