+++ docs/README.md
@@ -1,4 +1,4 @@
-# Pookie User Guide
+# pookie.Pookie User Guide
 
 // Update the title above to match the actual product name
