+++ src/main/java/pookie/CorruptFileException.java
@@ -1,3 +1,5 @@
+package pookie;
+
 public class CorruptFileException extends Exception {
     public CorruptFileException() {
         super("Save file is corrupt.");
