+++ src/main/java/pookie/TaskList.java
@@ -1,3 +1,7 @@
+package pookie;
+
+import pookie.model.Task;
+
 import java.util.ArrayList;
 
 public class TaskList {
