+++ src/main/java/pookie/command/ByeCommand.java
@@ -1,3 +1,9 @@
+package pookie.command;
+
+import pookie.Storage;
+import pookie.TaskList;
+import pookie.ui.Ui;
+
 public class ByeCommand extends Command {
     public void execute(String input, Ui ui, TaskList tasks, Storage storage, boolean testMode) throws Exception {
         ui.showMessage("Bye. Hope to see you again soon!");
