+++ src/main/java/pookie/command/DeadlineCommand.java
@@ -1,3 +1,9 @@
+package pookie.command;
+
+import pookie.*;
+import pookie.model.Deadline;
+import pookie.ui.Ui;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeParseException;
