+++ src/main/java/pookie/command/MarkCommand.java
@@ -1,22 +1,25 @@
-import java.io.IOException;
+package pookie.command;
 
-public class DeleteCommand extends Command {
+import pookie.Storage;
+import pookie.TaskList;
+import pookie.ui.Ui;
+
+public class MarkCommand extends Command {
     @Override
     public void execute(String input, Ui ui, TaskList tasks, Storage storage, boolean testMode) throws Exception {
         try {
             String[] parts = input.split(" ", 2);
             if (parts.length < 2) {
-                ui.showMessage("usage: delete <task number>");
+                ui.showMessage("usage: mark <task number>");
                 return;
             }
-
             int index = Integer.parseInt(parts[1].trim()) - 1;
+
             if (index >= 0 && index < tasks.size()) {
-                Task removedTask = tasks.remove(index);
+                tasks.get(index).markAsDone();
                 ui.showMessages(
-                    "Noted. I've removed this task:",
-                    "  " + removedTask,
-                    "Now you have " + tasks.size() + " tasks in the list."
+                    "Nice! I've marked this task as done:",
+                    "  " + tasks.get(index)
                 );
             } else {
                 ui.showInvalidTaskNumberError();
