+++ src/test/java/pookie/MockUi.java
@@ -1,3 +1,7 @@
+package pookie;
+
+import pookie.ui.Ui;
+
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
