+++ src/test/java/pookie/command/EventCommandTest.java
@@ -1,5 +1,9 @@
+package pookie.command;
+
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import pookie.MockUi;
+import pookie.TaskList;
 
 import java.util.ArrayList;
