+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
 ____________________________________________________________
- Hello! I'm Pookie
+ Hello! I'm pookie.Pookie
  What can I do for you?
 ____________________________________________________________
